--- Team1_Page/src/App.jsx
@@ -1,22 +1,0 @@
-import React from 'react';
-import { Routes, Route } from 'react-router-dom';
-import { HelmetProvider } from 'react-helmet-async';
-import { ThemeProvider, StyledEngineProvider } from '@mui/material/styles';
-import questTheme from 'src/QuestUiKitLightTheme';
-import TeamPage from 'src/components/TeamPage/TeamPage';
-
-function App() {
-  return (
-    <HelmetProvider>
-      <StyledEngineProvider injectFirst>
-        <ThemeProvider theme={questTheme}>
-          <Routes>
-            <Route path="/" element={<TeamPage />} />
-          </Routes>
-        </ThemeProvider>
-      </StyledEngineProvider>
-    </HelmetProvider>
-  );
-}
-
-export default App;
