+++ my-react-app/package.json
@@ -3,11 +3,22 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.10.4",
+    "@emotion/styled": "^11.10.4",
+    "@mui/icons-material": "^5.11.9",
+    "@mui/lab": "^5.0.0-alpha.102",
+    "@mui/material": "^5.10.5",
+    
+    "react-router-dom": "^6.4.0",
+    
+    "react-spring": "^9.5.5",
+    
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-helmet-async": "^1.3.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
